pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "magi-system"
version = "0.1.0"
description = "Disagreement OS for LLMs — three models, one decision"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Eric Chen" }]
dependencies = [
    "litellm>=1.0.0,!=1.82.7,!=1.82.8",
    "click>=8.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "websockets>=11.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["magi"]

[project.scripts]
magi = "magi.cli:main"
